[70218] Rename node image model class
