+++ Controller/Adminhtml/Menu/Save.php
@@ -14,7 +14,7 @@ use Snowdog\Menu\Api\Data\MenuInterface;
 use Snowdog\Menu\Api\MenuRepositoryInterface;
 use Snowdog\Menu\Api\NodeRepositoryInterface;
 use Snowdog\Menu\Model\Menu\NodeFactory;
-use Snowdog\Menu\Model\Menu\Node\Image as NodeImage;
+use Snowdog\Menu\Model\Menu\Node\Image\File as NodeImageFile;
 use Snowdog\Menu\Model\MenuFactory;
 use Snowdog\Menu\Service\MenuHydrator;
 
@@ -43,8 +43,8 @@ class Save extends Action
     /** @var MenuFactory */
     private $menuFactory;
 
-    /** @var NodeImage */
-    private $nodeImage;
+    /** @var NodeImageFile */
+    private $nodeImageFile;
 
     /** @var ProductRepository */
     private $productRepository;
@@ -61,7 +61,7 @@ class Save extends Action
      * @param SearchCriteriaBuilderFactory $searchCriteriaBuilderFactory
      * @param NodeFactory $nodeFactory
      * @param MenuFactory $menuFactory
-     * @param NodeImage $nodeImage
+     * @param NodeImageFile $nodeImageFile
      * @param ProductRepository $productRepository
      * @param MenuHydrator|null $hydrator
      */
@@ -74,7 +74,7 @@ class Save extends Action
         SearchCriteriaBuilderFactory $searchCriteriaBuilderFactory,
         NodeFactory $nodeFactory,
         MenuFactory $menuFactory,
-        NodeImage $nodeImage,
+        NodeImageFile $nodeImageFile,
         ProductRepository $productRepository,
         MenuHydrator $hydrator = null
     ) {
@@ -86,7 +86,7 @@ class Save extends Action
         $this->searchCriteriaBuilderFactory = $searchCriteriaBuilderFactory;
         $this->nodeFactory = $nodeFactory;
         $this->menuFactory = $menuFactory;
-        $this->nodeImage = $nodeImage;
+        $this->nodeImageFile = $nodeImageFile;
         $this->productRepository = $productRepository;
         // Backwards compatible class loader
         $this->hydrator = $hydrator ?? ObjectManager::getInstance()->get(MenuHydrator::class);
@@ -199,7 +199,7 @@ class Save extends Action
             $nodeObject->setPosition((string) $position);
 
             if ($nodeObject->getImage() && empty($node['image'])) {
-                $this->nodeImage->delete($nodeObject->getImage());
+                $this->nodeImageFile->delete($nodeObject->getImage());
             }
 
             $nodeObject->setImage($node['image'] ?? null);
